+++ src/routes/loginRoute.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import dotenv from "dotenv";
-import {addUser, findClient} from "../services/loginService.js";
+import {addUser, deleteUser, findUser, putUser} from "../services/loginService.js";
 
 dotenv.config({path: ".env"});
 
@@ -11,20 +11,42 @@ app.use(express.json());
 let router = express.Router();
 
 router.get('/login', (req, res) => {
-    console.log("GET request trial started.");
+    console.log("GET login request trial started.");
 
-    res.status(200).send({message: "Requisição GET OK"});
-    console.log("GET OK");
+    res.status(200).send({message: "Requisição GET login OK"});
     res.send({message: "GET OK"});
     return res.json("GET OK");
 });
 
-//Novo usuário (rota da página de login)
 router.post('/login', (req, res) => {
-    console.log("POST request trial started.");
+    console.log("POST login request trial started.");
 
     const {username, password} = req.body;
-    console.log(req.body);
+
+    if (!username || !password) {
+        res.status(400).send({
+            message: "Não foram recebidos os dados necessários."
+        });
+        console.log("Não foram recebidos os dados necessários.")
+        return res.json("Erro: não foram recebidos os dados.");
+    }
+
+    findUser(username, password)
+        .then(r => {
+            if (r) return res.status(200).json("Usuário encontrado.");
+            else return res.status(404).json("Não foi encontrado usuário.");
+        })
+        .catch(err => {
+            return res.status(500).json("Erro detectado: " + err.name)
+        })
+    ;
+
+});
+
+router.post('/signin', (req, res) => {
+    console.log("POST signin request trial started.");
+
+    const {username, password} = req.body;
 
     if (!username || !password) {
         res.status(400).send({
@@ -35,16 +57,17 @@ router.post('/login', (req, res) => {
     }
     console.log("Username recebido: " + username);
     console.log("Senha recebida: " + password);
-    addUser(username, password).then(() => res.json("Usuário adicionado."));  //Serviço
 
-    findClient(username, password).then(r => {return res.json("Usuário encontrado.")});
+    addUser(username, password)
+        .then(() => res.status(200).json("Usuário adicionado."))
+        .catch(err => {return res.status(500).json("Erro: " + err.name);})
+    ;
 });
 
-router.delete('/login', (req, res) => {
-    console.log("DELETE request trial started.");
+router.delete('/account', (req, res) => {
+    console.log("DELETE User request trial started.");
 
     const {username, password} = req.body;
-    console.log(req.body);
 
     if (!username || !password) {
         res.status(400).send({
@@ -54,7 +77,37 @@ router.delete('/login', (req, res) => {
         return res.json("Erro: não foram recebidos os dados.");
     }
 
-    //Incompleto
+    deleteUser(username, password)
+        .then(r => {
+            return res.json("Usuário deletado.");
+        })
+        .catch(err => {
+            return res.status(402).json("Erro detectado: " + err.name)
+        })
+    ;
+});
+
+router.put('/account', (req, res) => {
+    console.log("PUT User request trial started.");
+
+    const {username, password} = req.body;
+
+    if (!username || !password) {
+        res.status(400).send({
+            message: "Não foram recebidos os dados necessários."
+        });
+        console.log("Não foram recebidos os dados necessários.")
+        return res.json("Erro: não foram recebidos os dados.");
+    }
+
+    putUser(username, password)
+        .then(r => {
+            return res.status(200).json("Usuário atualizado.")
+        })
+        .catch(err => {
+            return res.status(404).json("Erro detectado: " + err.name)
+        })
+    ;
 });
 
 export default router;