+++ CHANGELOG.md
@@ -40,6 +40,7 @@ Aqui ocorrerá a listagem de mudanças no projeto e explicações ou consideraç
 34. Mudança na estrutura e no provedor para o deploy: [Render](https://render.com/)
 35. Criação do arquivo [API-LOG](API-LOG.md), para documentar a API
 36. Criação do arquivo [DB-LOG.md](DB-LOG.md) para explicar as definições sobre o banco de dados.
+37. Altarações no serviço e nas rotas de login
 
 ### Listagem de fontes ou referências de aprendizado